refactor(demo): extract DemoMCPCoordinator into its own file
Move ConnectedServicesCoordinator out of ConnectedServicesView.swift into
Example/BaseChatDemo/MCP/DemoMCPCoordinator.swift and rename it to DemoMCPCoordinator,
making the coordinator pattern easier for adopters to crib independently.
